moving spinner into re usable component lib
